% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-Metabase.R
\name{as_MicrobiomeSet}
\alias{as_MicrobiomeSet}
\title{Convert a phyloseq object to a MicrobiomeSet object}
\usage{
as_MicrobiomeSet(physeq)
}
\arguments{
\item{physeq}{\code{\link{phyloseq-class}} object}
}
\value{
\code{\link{MicrobiomeSet-class}} object
}
\description{
Convert a phyloseq object to a MicrobiomeSet object
}
\author{
Chenghao Zhu
}
