% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-Metabase.R
\name{as_MicrobiomeSetList}
\alias{as_MicrobiomeSetList}
\title{Convert a SummarizedPhyloseq object to a list of MicrobiomeSet}
\usage{
as_MicrobiomeSetList(spy)
}
\arguments{
\item{spy}{a \code{\link{SummarizedPhyloseq-class}} object}
}
\description{
Convert a SummarizedPhyloseq object to a list of MicrobiomeSet
}
\seealso{
\code{\link{phyloseq_to_mSet}} \code{\link{SummarizedPhyloseq-class}}
}
\author{
Chenghao Zhu
}
