% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeViz.R
\name{create_annodata}
\alias{create_annodata}
\title{Construct annotation table for microbiomeViz cladogram annotation}
\usage{
create_annodata(spys, coef = c("pvalue", "padj"), cutoff = 0.05,
  colors = c("#00468BFF", "#ED0000FF"), levels = 1:7)
}
\arguments{
\item{spys}{(required) A \code{\link{SummarizedPhyloseq-class}} object.}

\item{coef}{Must be pvalue or padj}

\item{cutoff}{The coef cutoff for selection. Default is 0.05}

\item{colors}{It takes a string vector of 2 colors. The first will be
positive logFC and the second will be negativ.}

\item{levels}{The phylogenic levels to annotate, with 1 being kingdom and
7 being species. Default is all levels.}
}
\description{
It creats a annotation table base on a SummarizedPhyloStats object with
node name and the color to highlight. This table can then be parsed to
the \code{\link{clade.anno}} function from the microbiomeViz package. It
goes through each slot in thee SummarizedPhyloStats object, and selects the
taxa that matches the provided criteria. For example, the default setting
selects all the taxa with a unadjusted pvalue small or equal to 0.05, and
label the ones with a positive logFC as red, and negative as blue.
}
\seealso{
\code{\link{SummarizedPhyloStats}}
\code{\link{clade.anno}}
}
