% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-fatigue.R
\docType{data}
\name{data-fatigue}
\alias{data-fatigue}
\alias{fatigue}
\title{(Data) Chronic Fatigue Syndrome dataset}
\description{
This dataset contains 87 individuals with 48 diseased patients with the
chronic fatigue syndrome, and 39 healthy controls. The data were sequenced on
an Illumina MiSeq using the Earth Microbiome Project hypervariable region 4
(V4) 16S rRNA sequencing protocol. The data processing procedure can be found
on the qiime2 documentation website.
}
\references{
Giloteaux, L., Goodrich, J. K., Walters, W. A., Levine, S. M.,
Ley, R. E., & Hanson, M. R. (2016).
Reduced diversity and altered composition of the gut microbiome in
individuals with myalgic encephalomyelitis/chronic fatigue syndrome.
Microbiome, 4(1), 30.
}
\author{
Guiloteaux et al.
}
\keyword{data}
