% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedPhyloseq-Assessor.R,
%   R/SummarizedPhyloStats-Assessor.R
\docType{methods}
\name{[[,SummarizedPhyloseq-method}
\alias{[[,SummarizedPhyloseq-method}
\alias{[[}
\alias{$,SummarizedPhyloseq-method}
\alias{$}
\alias{$<-,SummarizedPhyloseq-method}
\alias{$<-}
\alias{[[,SummarizedPhyloStats-method}
\alias{[[<-,SummarizedPhyloStats-method}
\alias{[[<-}
\alias{$,SummarizedPhyloStats-method}
\alias{$<-,SummarizedPhyloStats-method}
\title{Extract or replace a slot of an SummarizedPhyloseq or SummarizedPhyloStats object}
\usage{
\S4method{[[}{SummarizedPhyloseq}(x, i, j, drop)

\S4method{$}{SummarizedPhyloseq}(x, name)

\S4method{$}{SummarizedPhyloseq}(x, name) <- value

\S4method{[[}{SummarizedPhyloStats}(x, i, j, drop)

\S4method{[[}{SummarizedPhyloStats}(x, i, j, ...) <- value

\S4method{$}{SummarizedPhyloStats}(x, name)

\S4method{$}{SummarizedPhyloStats}(x, name) <- value
}
\arguments{
\item{x}{An \code{\link{SummarizedPhyloseq-class}} or \code{\link{SummarizedPhyloStats-class}} object.}

\item{name}{The name of the slot.}
}
\description{
Extract or replace the content of a slot of an
\code{\link{SummarizedPhyloseq-class}} or an \code{\link{SummarizedPhyloStats-class}} object.
}
