% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-plot.R
\name{plot_bar}
\alias{plot_bar}
\title{Stacked bar plot for SummarizedPhyloseq data}
\usage{
plot_bar(spy, level = "Phylum", by = NULL, show.legend = TRUE)
}
\arguments{
\item{spy}{\code{\link{SummarizedPhyloseq-class}}}

\item{level}{Character. The phyloseq level to use. Must be one of
Kingdom, Phylum, Class, Order, Family, Genus, or Species.}

\item{by}{Character. The sample meta-data variable to plot the
abundance data againt to. It must be from the sample data's column names.
If multiple variables are givin, facets will be used. The default is
sample name. See examples.}

\item{legend.show}{Logical variable whether to show the legends.}

\item{plotly}{Logical value. If TRUE, a plotly variable will be returned.}
}
\description{
This function uses the ggplot2 package to generate barplot from a given
\code{\link{SummarizedPhyloseq-class}} object on a given phylogenic level.
The output is a ggplot object so it can be easily styled using additional
ggplot functions.
}
\examples{
data(fatigue)
fatigue = transform_sample_counts(physeq, function(x) x/sum(x))
spy = summarizeFromPhyloseq(fatigue)
plot_bar(spy)
plot_bar(spy, "Phylum", by = "Subject")
plot_bar(spy, "Genus", by = c("Subject", "Sex"))
plot_bar(spy, "Genus", by = c("Subject", "Sex"), plotly = T)
plot_bar(spy, "Phylum", by = c("Subject", "age_range", "Sex"))
}
\author{
Chenghao Zhu
}
