% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_to_stats.R
\name{spy_to_deseq2}
\alias{spy_to_deseq2}
\title{Conduct DESeq2 analysis on different phylogenic levels}
\usage{
spy_to_deseq2(spy, design, resultsName, ...)
}
\arguments{
\item{spy}{(required) \code{\link{SummarizedPhyloseq-class}}}

\item{design}{(required) The design matrix of the experiment, with rows
corresponding to sample IDs and columns to coefficients to be estimated.
This can be constructed using \code{\link{model.matrix}} function}

\item{resultsName}{The index or name of the coefficient (variable) for
building the results tables. The value provided must be one of the colnames
of the design matrix. This argument is corresponding to the \code{name} in
the \code{\link{results}} from the DESeq2 package.}

\item{...}{(optional) Other parameters to be parsed to the
\code{\link{DESeq}} function}
}
\description{
Conduct differential expression using the DESeq2 package to a
SummarizedPhyloseq object. See the \code{\link{DESeq}} for more detail.
}
\note{
The DESeq2 package needs to be installed in order to successfully run
this function.
}
\author{
Chenghao Zhu
}
