% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedPhyloseq-Assessor.R
\docType{methods}
\name{[[<-,SummarizedPhyloseq-method}
\alias{[[<-,SummarizedPhyloseq-method}
\alias{[[<-}
\title{rdname Extract}
\usage{
\S4method{[[}{SummarizedPhyloseq}(x, i, j, ...) <- value
}
\description{
rdname Extract
}
