% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_taxa}
\alias{summarize_taxa}
\title{summarize_taxa}
\usage{
summarize_taxa(physeq, level = "Family", keep_full_tax = TRUE)
}
\arguments{
\item{physeq}{a phyloseq object}

\item{level}{the taxonomy level to summarize. Level must be one from 1 to 7, while 1 is Kingdom and 7 is Species.}

\item{keep_full_tax}{logistical value whether to use the full tax path. Default is FALSE.}
}
\description{
summarize phyloseq object on different taxonomy level.
}
\examples{
library(phyloseq)
data("GlobalPatterns")
GP_family = summarize_taxa(GlobalPatterns, level=5)
}
\author{
Chenghao Zhu
}
