% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedPhyloseq-Assessor.R,
%   R/SummarizedPhyloStats-Assessor.R
\docType{methods}
\name{kingdom_table}
\alias{kingdom_table}
\alias{phylum_table,SummarizedPhyloseq-method}
\alias{phylum_table}
\alias{class_table,SummarizedPhyloseq-method}
\alias{class_table}
\alias{order_table,SummarizedPhyloseq-method}
\alias{order_table}
\alias{family_table,SummarizedPhyloseq-method}
\alias{family_table}
\alias{genus_table,SummarizedPhyloseq-method}
\alias{genus_table}
\alias{species_table,SummarizedPhyloseq-method}
\alias{species_table}
\alias{kingdom_table,SummarizedPhyloStats-method}
\alias{phylum_table,SummarizedPhyloStats-method}
\alias{class_table,SummarizedPhyloStats-method}
\alias{order_table,SummarizedPhyloStats-method}
\alias{family_table,SummarizedPhyloStats-method}
\alias{genus_table,SummarizedPhyloStats-method}
\alias{species_table,SummarizedPhyloStats-method}
\alias{otu_table,SummarizedPhyloStats-method}
\title{Accessors for SummarizedPhyloseq class and SummarizedPhyloStats class}
\usage{
\S4method{kingdom_table}{SummarizedPhyloseq}(object)

\S4method{phylum_table}{SummarizedPhyloseq}(object)

\S4method{class_table}{SummarizedPhyloseq}(object)

\S4method{order_table}{SummarizedPhyloseq}(object)

\S4method{family_table}{SummarizedPhyloseq}(object)

\S4method{genus_table}{SummarizedPhyloseq}(object)

\S4method{species_table}{SummarizedPhyloseq}(object)

\S4method{kingdom_table}{SummarizedPhyloStats}(object)

\S4method{phylum_table}{SummarizedPhyloStats}(object)

\S4method{class_table}{SummarizedPhyloStats}(object)

\S4method{order_table}{SummarizedPhyloStats}(object)

\S4method{family_table}{SummarizedPhyloStats}(object)

\S4method{genus_table}{SummarizedPhyloStats}(object)

\S4method{species_table}{SummarizedPhyloStats}(object)

\S4method{otu_table}{SummarizedPhyloStats}(object)
}
\arguments{
\item{object}{Either a SummarizedPhyloseq or SummarziedPhyloStats object.}
}
\value{
A otu_table-class object.
}
\description{
Accessor methods for the \code{\link{SummarizedPhyloseq-class}}
and the \code{\link{SummarizedPhyloStats-class}}
}
\details{
These methods retrieve slots from either a SummarizedPhyloseq or a SummarizedPhyloStats object.
\describe{
    \item{kingdom_table()}{Retrieve the kingdom_table slot}
    \item{phylum_table()}{Retrieve the phylum_table slot}
    \item{class_table()}{Retrieve the class_table slot}
    \item{order_table()}{Retrieve the order_table slot}
    \item{family_table()}{Retrieve the family_table slot}
    \item{genus_table()}{Retrieve the genus_table slot}
    \item{species_table()}{Retrieve the species_table slot}
    \item{otu_table()}{Retrieve the otu_table slot}
}
}
\seealso{
\code{\link{SummarizedPhyloseq-class}}
\code{\link{SummarizedPhyloStats-class}}
\code{\link{summarizeFromPhyloseq}}
}
\author{
Chenghao Zhu
}
