% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedPhyloseq-Assignment.R,
%   R/SummarizedPhyloStats-Assignment.R
\docType{methods}
\name{kingdom_table<-}
\alias{kingdom_table<-}
\alias{phylum_table<-,SummarizedPhyloseq-method}
\alias{phylum_table<-}
\alias{class_table<-,SummarizedPhyloseq-method}
\alias{class_table<-}
\alias{order_table<-,SummarizedPhyloseq-method}
\alias{order_table<-}
\alias{family_table<-,SummarizedPhyloseq-method}
\alias{family_table<-}
\alias{genus_table<-,SummarizedPhyloseq-method}
\alias{genus_table<-}
\alias{species_table<-,SummarizedPhyloseq-method}
\alias{species_table<-}
\alias{kingdom_table<-,SummarizedPhyloStats-method}
\alias{kingdom_table<-}
\alias{phylum_table<-,SummarizedPhyloStats-method}
\alias{phylum_table<-}
\alias{class_table<-,SummarizedPhyloStats-method}
\alias{calss_table<-}
\alias{order_table<-,SummarizedPhyloStats-method}
\alias{order_table<-}
\alias{family_table<-,SummarizedPhyloStats-method}
\alias{family_table<-}
\alias{genus_table<-,SummarizedPhyloStats-method}
\alias{genus_table<-}
\alias{species_table<-,SummarizedPhyloStats-method}
\alias{species_table<-}
\alias{otu_table<-,SummarizedPhyloStats,ANY-method}
\title{Assignment methods for SummarizedPhyloseq and SummarizedPhyloStats}
\usage{
\S4method{kingdom_table}{SummarizedPhyloseq}(object) <- value

\S4method{phylum_table}{SummarizedPhyloseq}(object) <- value

\S4method{class_table}{SummarizedPhyloseq}(object) <- value

\S4method{order_table}{SummarizedPhyloseq}(object) <- value

\S4method{family_table}{SummarizedPhyloseq}(object) <- value

\S4method{genus_table}{SummarizedPhyloseq}(object) <- value

\S4method{species_table}{SummarizedPhyloseq}(object) <- value

\S4method{kingdom_table}{SummarizedPhyloStats}(object) <- value

\S4method{phylum_table}{SummarizedPhyloStats}(object) <- value

\S4method{class_table}{SummarizedPhyloStats}(object) <- value

\S4method{order_table}{SummarizedPhyloStats}(object) <- value

\S4method{family_table}{SummarizedPhyloStats}(object) <- value

\S4method{genus_table}{SummarizedPhyloStats}(object) <- value

\S4method{species_table}{SummarizedPhyloStats}(object) <- value

\S4method{otu_table}{SummarizedPhyloStats,ANY}(x) <- value
}
\arguments{
\item{object}{(Required) \code{\link{SummarizedPhyloseq-class}} or \code{\link{SummarizedPhyloStats-class}}}

\item{value}{(Required) \code{\link{otu_table-class}}}
}
\description{
Assignment methods for the \code{\link{SummarizedPhyloseq-class}}
class and the \code{\link{SummarizedPhyloStats-class}} class
}
\details{
These methods assign an new \code{\link{otu_table-class}} object to either a SummarizedPhyloseq or a SummarizedPhyloStats object.
}
\seealso{
\code{\link{SummarizedPhyloseq-class}}
\code{\link{SummarizedPhyloStats-class}}
\code{\link{summarizeFromPhyloseq}}
}
