% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedPhyloseq-Class.R
\docType{class}
\name{SummarizedPhyloseq-class}
\alias{SummarizedPhyloseq-class}
\title{S4 class SummarizedPhyloseq}
\description{
A S4 Object for storing phyloseq data at all phylogenic levels
}
\details{
This class inherit from the
\code{\link{phyloseq-class}}
from the phyloseq package. It contains three data classes:
\code{\link{otu_table-class}} ("otu_table" slot),
\code{\link{sample_data-class}} ("sample_data" slot), and
\code{\link{taxonomyTable-class}} ("tax_table" slot).
Besides the three slots inherit from phyloseq-class, the SummarizedPhyloseq
also has 7 more slots, "kingdom_table", "phylum_table", "class_table",
"order_table", "family_table", "genus_table", and "species_table", all in
the class of otu_table-class. This allows you to view the data in different
phylogenic levels.
\code{\link{summarizeFromPhyloseq}}
is the main constructor for this class, directly from a phyloseq object.
The phyloseq needs to have the otu_table, tax_table, and sample_data to
successfully construct this class.

slots:
\describe{
    \item{otu_table}{ a tabular object of the OTU table, class of otu_table-class}
    \item{sam_data}{ a tabular object of the sample meta-data, class of sample_data.}
    \item{tax_table}{ a tabular object of the taxonomy table, class of taxonomyTable.}
    \item{kingdom_table}{ a tabular object of the summarized OTU table into the kingdom level, class of otu_table}
    \item{phylum_table}{ a tabular object of the summarized OTU table into the phylum level, class of otu_table}
    \item{class_table}{ a tabular object of the summarized OTU table into the class level, class of otu_table}
    \item{order_table}{ a tabular object of the summarized OTU table into the order level, class of otu_table}
    \item{family_table}{ a tabular object of the summarized OTU table into the family level, class of otu_table}
    \item{genus_table}{ a tabular object of the summarized OTU table into the genus level, class of otu_table}
    \item{species_table}{ a tabular object of the summarized OTU table into the species level, class of otu_table}
}
}
\seealso{
\code{\link{phyloseq-class}}
\code{\link{summarizeFromPhyloseq}}
\code{\link{SummarizedPhyloStats-class}}
\code{\link{SummarizedPhyloseq-accessors}}
\code{\link{SummarizedPhyloseq-assign}}
}
\author{
Chenghao Zhu
}
