% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedPhyloStats-Class.R
\docType{class}
\name{SummarizedPhyloStats-class}
\alias{SummarizedPhyloStats-class}
\title{S4 class SummarizedPhyloStats}
\description{
A S4 object to store hypothesis test statistical results of phylogenic
sequencing experiment data, at different phylogenic level
}
\details{
This class has 8 slots and each stores the hypothesis test statistical
results of a designated phylogenic level, from otu to kingdom.There are
currently two statistical packages supported, limma and DESeq2. There are
2 construction methods \code{\link{spy_to_deseq2}} and
\code{\link{spy_to_limma}}. The purpose of this class is to allow users to
conduct hypothesis test using DESeq2 or limma at all phylogenic levels and
view the results quickly.

\describe{
    This class has 8 slots and each slot is a data.frame with 5 variables as
below:
    \item{baseMean}{The mean value of the baseline group. This corresponds
    to the AveExpr from the limma result or the baseMean
    from the DESeq2 result}
    \item{logFC}{Log transformed fold change from baseline. This corresponds
    to the logFC from the limma result or the log2FoldChange from the DESeq2
    result}
    \item{stat}{The statistic value. This corresponds to the t from the limma
    result or the stat from the DESeq2 result}
    \item{pvalue}{The raw p-value. This corresponds to the P.Value from the
    limma result or the pvalue from the DESeq2 result}
    \item{padj}{The adjusted p-value. This corresponds to the adj.P.Value
    from the limma result or the padj from the DESeq2 result}
}
}
\section{Slots}{

\describe{
\item{\code{kingdom_table}}{Summarized statistic result at kingdom level.}

\item{\code{phylum_table}}{Summarized statistic result at phylum level.}

\item{\code{class_table}}{Summarized statistic result at class level.}

\item{\code{order_table}}{Summarized statistic result at order level.}

\item{\code{family_table}}{Summarized statistic result at family level.}

\item{\code{genus_table}}{Summarized statistic result at genus level.}

\item{\code{species_table}}{Summarized statistic result at species level.}

\item{\code{otu_table}}{Summarized statistic result at otu level.}
}}

\seealso{
\code{\link{spy_to_deseq2}}
\code{\link{spy_to_limma}}
\code{\link{summarizeFromPhyloseq}}
\code{\link{SummarizedPhyloseq-accessors}}
\code{\link{SummarizedPhyloseq-assign}}
\code{\link{phyloseq-class}}
}
\author{
Chenghao Zhu
}
