% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizedPhyloseq-Class.R
\name{summarizeFromPhyloseq}
\alias{summarizeFromPhyloseq}
\title{summarizeFromPhyloseq}
\usage{
summarizeFromPhyloseq(physeq, level = "all", keep_full_tax = FALSE)
}
\arguments{
\item{physeq}{An phyloseq object from the phyloseq package. It must contain
an otu_table slot, a sam_data slot, and a tax_table slot. The phy_tree slot
and refseq slots are not required.}

\item{level}{The taxonomy level to summarize. It can be one ore more from
Kingdom, Phylum, Class, Order, Family, Genus, and Species. If mutiple
phylogenic levels are specified, use a string vector. Default is "all" that
parses all levels. The slots that are not specified will be NULL.}
}
\value{
A SummarizedPhyloseq object
}
\description{
Construct a
\code{\link{SummarizedPhyloseq-class}} object from a
\code{\link{phyloseq-class}} object
}
\details{
This is the main constructor of the SummarizedPhyloseq class. It takes a
phyloseq object and calls the
\code{link{summarize_taxa}}() function for each phylogenic level and return
a SummarizedPhyloseq object.
}
\examples{
library(phyloseq)
data(GlobalPatterns)
spy = summarizedFromPhyloseq(GlobalPatterns)
}
\seealso{
\code{\link{SummarizedPhyloseq-class}}
}
\author{
Chenghao Zhu
}
